Watchdog - improve memory footprint, uint8_t are enough to store flags.

+++ include/core/watchdog.h
@@ -42,8 +42,8 @@
 #define WDT_CLK_WDTCLK 1
 struct wdt_config {
  	/* clk_sel is either 0 (IRC) or 1 (WDTCLK). The corresponding clock source will be powered on. */
-	int clk_sel;
-	int intr_mode_only; /* If set to 1, a watchdog timeout will trigger an interrupt instead of a reset */
+	uint8_t clk_sel;
+	uint8_t intr_mode_only; /* If set to 1, a watchdog timeout will trigger an interrupt instead of a reset */
 	void (*callback)(void);
 	uint32_t locks; /* Bitfield from WDT_*_LOCK defined in watchdog.h */
 	/* Number of clk_src clocks before the watchdog timer times out. Will be divided by 4 to give
